+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/add/GroupAddViewModel.kt
@@ -2,5 +2,6 @@ package com.woowacourse.friendogly.presentation.ui.group.add
 
 import com.woowacourse.friendogly.presentation.base.BaseViewModel
 
-class GroupAddViewModel: BaseViewModel() {}
+class GroupAddViewModel: BaseViewModel() {
 
+}
