+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/list/adapter/filter/FilterViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.friendogly.presentation.ui.group.list.adapter.filter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.woowacourse.friendogly.databinding.ItemFilterBinding
-import com.woowacourse.friendogly.presentation.ui.group.list.model.groupfilter.GroupFilter
+import com.woowacourse.friendogly.presentation.ui.group.model.groupfilter.GroupFilter
 
 class FilterViewHolder(
     private val binding: ItemFilterBinding,
