+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/list/adapter/group/GroupListAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.woowacourse.friendogly.databinding.ItemGroupBinding
 import com.woowacourse.friendogly.presentation.ui.group.list.GroupListActionHandler
-import com.woowacourse.friendogly.presentation.ui.group.list.model.GroupUiModel
+import com.woowacourse.friendogly.presentation.ui.group.model.GroupUiModel
 
 class GroupListAdapter(
     private val actionHandler: GroupListActionHandler,
