+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/model/GroupFilterSelector.kt
@@ -1,12 +1,13 @@
-package com.woowacourse.friendogly.presentation.ui.group.list.model
+package com.woowacourse.friendogly.presentation.ui.group.model
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.woowacourse.friendogly.presentation.ui.group.list.model.groupfilter.GroupFilter
-import com.woowacourse.friendogly.presentation.ui.group.list.model.groupfilter.ParticipationFilter
+import com.woowacourse.friendogly.presentation.ui.group.model.groupfilter.GroupFilter
+import com.woowacourse.friendogly.presentation.ui.group.model.groupfilter.ParticipationFilter
 
 class GroupFilterSelector {
-    private val _participationFilter: MutableLiveData<ParticipationFilter> = MutableLiveData(ParticipationFilter.POSSIBLE)
+    private val _participationFilter: MutableLiveData<ParticipationFilter> = MutableLiveData(
+        ParticipationFilter.POSSIBLE)
     val participationFilter: LiveData<ParticipationFilter> get() = _participationFilter
 
     private val _currentSelectedFilters: MutableLiveData<List<GroupFilter>> = MutableLiveData()
