+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/model/GroupUiModel.kt
@@ -1,6 +1,6 @@
-package com.woowacourse.friendogly.presentation.ui.group.list.model
+package com.woowacourse.friendogly.presentation.ui.group.model
 
-import com.woowacourse.friendogly.presentation.ui.group.list.model.groupfilter.GroupFilter
+import com.woowacourse.friendogly.presentation.ui.group.model.groupfilter.GroupFilter
 import java.time.LocalDateTime
 
 data class GroupUiModel(
