+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/model/GroupWoof.kt
