+++ android/app/src/main/res/layout/fragment_group_add.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.woowacourse.friendogly.presentation.ui.group.list.model.groupfilter.GroupFilter" />
+        <import type="com.woowacourse.friendogly.presentation.ui.group.model.groupfilter.GroupFilter" />
 
         <variable
             name="vm"
